Rename export/import-snapshot commands to export/import-archive
The database file export/import feature is now called "archive" instead of
"snapshot" to avoid confusion with the API-based snapshot commands. The old
export-snapshot and import-snapshot commands are kept as hidden deprecated
aliases that warn users to switch to the new names.
